fix: backhandler remove old api
