+++ packages/unstyled/hooks/src/use-keyboard-dismissable/index.ts
@@ -1,6 +1,5 @@
 import React from 'react';
-import { useEffect } from 'react';
-import { BackHandler, Platform } from 'react-native';
+import { useBackHandler } from '../use-back-handler';
 
 type IParams = {
   enabled?: boolean;
@@ -41,32 +40,3 @@ export const useKeyboardDismissable = ({ enabled, callback }: IParams) => {
 
   useBackHandler({ enabled, callback });
 };
-
-export function useBackHandler({ enabled, callback }: IParams) {
-  useEffect(() => {
-    if (Platform.OS === 'web') {
-      const handleEscape = (e: KeyboardEvent) => {
-        if (e.key === 'Escape') {
-          callback();
-        }
-      };
-
-      document?.body?.addEventListener?.('keyup', handleEscape);
-      return () => {
-        document?.body?.removeEventListener?.('keyup', handleEscape);
-      };
-    } else {
-      const backHandler = () => {
-        callback();
-        return true;
-      };
-      if (enabled) {
-        BackHandler.addEventListener('hardwareBackPress', backHandler);
-      } else {
-        BackHandler.removeEventListener('hardwareBackPress', backHandler);
-      }
-      return () =>
-        BackHandler.removeEventListener('hardwareBackPress', backHandler);
-    }
-  }, [enabled, callback]);
-}
